+++ package.json
@@ -48,6 +48,7 @@
     "react-chartjs-2": "^4.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18",
+    "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.55.0",
     "react-lottie-player": "^2.1.0",
     "react-quill-new": "^3.4.6",
